clean up the lowerin
