+++ protocols/src/ir/mod.rs
@@ -4,6 +4,7 @@ pub mod graph_interpreter;
 pub mod graphviz;
 pub mod lowering;
 pub mod proto_graph;
+pub mod trace_lowering;
 
 // TODO: add a function to transform AST to IR
 // pub fn frontend(
